+++ Sources/ProjectSpec/Scheme.swift
@@ -184,6 +184,7 @@ public struct Scheme: Equatable {
         public var deleteScreenshotsWhenEachTestSucceeds: Bool
 
         public struct TestTarget: Equatable, ExpressibleByStringLiteral {
+            
             public static let randomExecutionOrderDefault = false
             public static let parallelizableDefault = false
 
@@ -191,6 +192,7 @@ public struct Scheme: Equatable {
             public let targetReference: TargetReference
             public var randomExecutionOrder: Bool
             public var parallelizable: Bool
+            public var location: String?
             public var skipped: Bool
             public var skippedTests: [String]
             public var selectedTests: [String]
@@ -199,6 +201,7 @@ public struct Scheme: Equatable {
                 targetReference: TargetReference,
                 randomExecutionOrder: Bool = randomExecutionOrderDefault,
                 parallelizable: Bool = parallelizableDefault,
+                location: String? = nil,
                 skipped: Bool = false,
                 skippedTests: [String] = [],
                 selectedTests: [String] = []
@@ -206,6 +209,7 @@ public struct Scheme: Equatable {
                 self.targetReference = targetReference
                 self.randomExecutionOrder = randomExecutionOrder
                 self.parallelizable = parallelizable
+                self.location = location
                 self.skipped = skipped
                 self.skippedTests = skippedTests
                 self.selectedTests = selectedTests
@@ -216,6 +220,7 @@ public struct Scheme: Equatable {
                     targetReference = try TargetReference(value)
                     randomExecutionOrder = false
                     parallelizable = false
+                    location = nil
                     skipped = false
                     skippedTests = []
                     selectedTests = []
@@ -536,6 +541,7 @@ extension Scheme.Test.TestTarget: JSONObjectConvertible {
         targetReference = try TargetReference(jsonDictionary.json(atKeyPath: "name"))
         randomExecutionOrder = jsonDictionary.json(atKeyPath: "randomExecutionOrder") ?? Scheme.Test.TestTarget.randomExecutionOrderDefault
         parallelizable = jsonDictionary.json(atKeyPath: "parallelizable") ?? Scheme.Test.TestTarget.parallelizableDefault
+        location = jsonDictionary.json(atKeyPath: "location") ?? nil
         skipped = jsonDictionary.json(atKeyPath: "skipped") ?? false
         skippedTests = jsonDictionary.json(atKeyPath: "skippedTests") ?? []
         selectedTests = jsonDictionary.json(atKeyPath: "selectedTests") ?? []
@@ -558,6 +564,9 @@ extension Scheme.Test.TestTarget: JSONEncodable {
         }
         if parallelizable != Scheme.Test.TestTarget.parallelizableDefault {
             dict["parallelizable"] = parallelizable
+        }
+        if let location = location {
+            dict["location"] = location
         }
         if skipped {
             dict["skipped"] = skipped